Refactor socket route and add new socketio API endpoint
